requirements.txt
tensorflow==2.10.1
numpy<1.24
protobuf==3.19.6
pandas>=1.5
scikit-learn>=1.2
pillow>=9.4
streamlit==1.17.0
altair<5
